OTT-1969 :: Adpod request processing framework
